+++ src/results-output/output.ts
@@ -7,6 +7,12 @@ import * as cursorUtil from '../cursor-doc/utilities';
 import * as chalk from 'chalk';
 import * as ansiRegex from 'ansi-regex';
 import * as printer from '../printer';
+import {
+  appendToReplOutputWebview,
+  showReplOutputWebviewPanel,
+  appendStackTraceToReplOutputWebview,
+} from '../../out/cljs-lib/cljs-lib';
+import * as replSession from '../nrepl/repl-session';
 
 const customChalk = new chalk.Instance({ level: 3 });
 
@@ -79,7 +85,7 @@ export interface AfterAppendCallback {
   (insertLocation: vscode.Location, newPosition?: vscode.Location): any;
 }
 
-export type OutputDestination = 'repl-window' | 'output-channel' | 'terminal';
+export type OutputDestination = 'repl-window' | 'output-channel' | 'terminal' | 'output-view';
 
 export type OutputDestinationConfiguration = {
   evalResults: OutputDestination;
@@ -162,6 +168,9 @@ export function showResultOutputDestination(preserveFocus = true) {
   if (getDestinationConfiguration().evalResults === 'terminal') {
     return showOutputTerminal(preserveFocus);
   }
+  if (getDestinationConfiguration().evalResults === 'output-view') {
+    return showReplOutputWebviewPanel(preserveFocus);
+  }
   return outputWindow.revealResultsDoc(preserveFocus);
 }
 
@@ -182,11 +191,12 @@ function messageContainsAnsi(message: string) {
 }
 
 // Used to decide if new result output should be prepended with a newline or not.
-// Also: For non-result output, whether the repl window output should be be printed as line comments.
+// Also: For non-result output, whether the repl window output should be printed as line comments.
 const didLastOutputTerminateLine: Record<OutputDestination, boolean> = {
   'repl-window': true,
   'output-channel': true,
   terminal: true,
+  'output-view': true,
 };
 
 let havePrintedLegacyReplWindowOutputMessage = false;
@@ -208,6 +218,7 @@ const lastInfoLineData: Record<OutputDestination, AppendClojureOptions> = {
   'repl-window': {},
   'output-channel': {},
   terminal: {},
+  'output-view': {},
 };
 
 function saveLastInfoLineData(destination: OutputDestination, options: AppendClojureOptions) {
@@ -262,6 +273,11 @@ function appendClojure(
     getOutputPTY().write(`${didLastTerminateLine ? '' : '\n'}${nsInfoLine(destination, options)}`);
     // getOutputPTY().write(`${didLastTerminateLine ? '' : '\n'}`);
     getOutputPTY().write(`${prettyMessage}\n`);
+    if (after) {
+      after(undefined, undefined);
+    }
+  } else if (destination === 'output-view') {
+    appendToReplOutputWebview(options, message);
     if (after) {
       after(undefined, undefined);
     }
@@ -330,6 +346,9 @@ function append(options: AppendOptions, message: string, after?: AfterAppendCall
       after(undefined, undefined);
     }
     return;
+  }
+  if (destination === 'output-view') {
+    appendToReplOutputWebview(options, message);
   }
 }
 
@@ -421,6 +440,9 @@ function appendLine(options: AppendOptions, message: string, after?: AfterAppend
   if (destination === 'terminal') {
     append(options, message + '\r\n', after);
   }
+  if (destination === 'output-view') {
+    appendToReplOutputWebview(options, message);
+  }
 }
 
 /**
@@ -495,4 +517,59 @@ export function appendLineOtherErr(message: string, after?: AfterAppendCallback)
 export function replWindowAppendPrompt(onAppended?: outputWindow.OnAppendedCallback) {
   didLastOutputTerminateLine['output-window'] = true;
   outputWindow.appendPrompt(onAppended);
+}
+
+function formatStacktrace(stacktrace: any[]) {
+  return stacktrace
+    .filter((entry) => {
+      return (
+        !entry.flags.includes('dup') &&
+        !['clojure.lang.RestFn', 'clojure.lang.AFn'].includes(entry.class)
+      );
+    })
+    .map((entry) => {
+      const name = entry.var || entry.name;
+      return `${name} (${entry.file}:${entry.line})`;
+    })
+    .join('\n');
+}
+
+function printStackTrace(stacktrace: any[]) {
+  const evalResultsOutputDestination = getDestinationConfiguration().evalResults;
+  switch (evalResultsOutputDestination) {
+    case 'repl-window':
+      outputWindow.printLastStacktrace();
+      replWindowAppendPrompt();
+      break;
+    case 'output-view':
+      appendStackTraceToReplOutputWebview(stacktrace);
+      break;
+    case 'output-channel':
+      outputChannel.appendLine('');
+      outputChannel.appendLine(formatStacktrace(stacktrace));
+      break;
+    case 'terminal':
+      getOutputPTY().write('\n' + formatStacktrace(stacktrace) + '\n');
+      break;
+    default:
+      console.error(
+        'Printing the last stacktrace is not supported for the configured results output destination:',
+        evalResultsOutputDestination
+      );
+      break;
+  }
+}
+
+export function printLastStacktrace() {
+  const session = replSession.getSession();
+  session
+    .stacktrace()
+    .then((stacktrace) => {
+      if (stacktrace.stacktrace) {
+        printStackTrace(stacktrace.stacktrace);
+      }
+    })
+    .catch((e) => {
+      console.error(`Failed fetching stacktrace: ${e.message}`);
+    });
 }